backend/pyproject.toml
[project]
name = "medha-algo-pilot"
version = "1.0.0"
description = "Production-grade algorithmic trading platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "alembic==1.12.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "httpx==0.25.1",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "email-validator>=2.2.0",
    "celery==5.3.4",
    "redis==5.0.1",
    "structlog==23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include all Python packages in the backend directory
only-include = [
    "api",
    "database", 
    "trading_engine",
    "risk_engine",
    "broker",
    "strategies",
    "utils",
    "config.py",
    "run.py"
]

